refactor(s3): split s3 types and ops crates
