+++ Cargo.toml
@@ -39,7 +39,8 @@ members = [
     "crates/protos", # Protocol buffer definitions
     "crates/rio", # Rust I/O utilities and abstractions
     "crates/concurrency", # Concurrency management for RustFS - timeout, locking, backpressure, and I/O scheduling
-    "crates/s3-common", # Common utilities and data structures for S3 compatibility
+    "crates/s3-types", # S3 event type definitions
+    "crates/s3-ops", # S3 operation definitions and mapping
     "crates/s3select-api", # S3 Select API interface
     "crates/s3select-query", # S3 Select query engine
     "crates/scanner", # Scanner for data integrity checks and health monitoring
@@ -101,7 +102,8 @@ rustfs-policy = { path = "crates/policy", version = "1.0.0-beta.3" }
 rustfs-protos = { path = "crates/protos", version = "1.0.0-beta.3" }
 rustfs-protocols = { path = "crates/protocols", version = "1.0.0-beta.3" }
 rustfs-rio = { path = "crates/rio", version = "1.0.0-beta.3" }
-rustfs-s3-common = { path = "crates/s3-common", version = "1.0.0-beta.3" }
+rustfs-s3-types = { path = "crates/s3-types", version = "1.0.0-beta.3" }
+rustfs-s3-ops = { path = "crates/s3-ops", version = "1.0.0-beta.3" }
 rustfs-s3select-api = { path = "crates/s3select-api", version = "1.0.0-beta.3" }
 rustfs-s3select-query = { path = "crates/s3select-query", version = "1.0.0-beta.3" }
 rustfs-scanner = { path = "crates/scanner", version = "1.0.0-beta.3" }
