+++ crates/notify/examples/full_demo.rs
@@ -24,7 +24,7 @@ use rustfs_config::{
 use rustfs_ecstore::config::{Config, KV, KVS};
 use rustfs_notify::{BucketNotificationConfig, Event, NotificationError};
 use rustfs_notify::{initialize, notification_system};
-use rustfs_s3_common::EventName;
+use rustfs_s3_types::EventName;
 use rustfs_targets::arn::TargetID;
 use std::sync::Arc;
 use std::time::Duration;
