+++ crates/notify/src/global.rs
@@ -17,7 +17,7 @@ use crate::{
     NotificationSystem, NotificationTargetMetricSnapshot,
 };
 use rustfs_ecstore::config::Config;
-use rustfs_s3_common::EventName;
+use rustfs_s3_types::EventName;
 use rustfs_targets::arn::TargetID;
 use std::sync::{Arc, OnceLock};
 use tracing::error;
