+++ crates/notify/src/rule_engine.rs
@@ -14,7 +14,7 @@
 
 use crate::rules::{RulesMap, TargetIdSet};
 use percent_encoding::percent_decode_str;
-use rustfs_s3_common::EventName;
+use rustfs_s3_types::EventName;
 use rustfs_targets::arn::TargetID;
 use starshard::{AsyncShardedHashMap, DEFAULT_SHARDS, SnapshotMode};
 use std::sync::Arc;
@@ -99,7 +99,7 @@ impl Default for NotifyRuleEngine {
 mod tests {
     use super::NotifyRuleEngine;
     use crate::rules::RulesMap;
-    use rustfs_s3_common::EventName;
+    use rustfs_s3_types::EventName;
     use rustfs_targets::arn::TargetID;
 
     #[tokio::test]
