+++ crates/notify/src/rules/pattern_rules_test.rs
@@ -19,7 +19,7 @@
 //! configuration to event matching.
 
 use super::*;
-use rustfs_s3_common::EventName;
+use rustfs_s3_types::EventName;
 use rustfs_targets::arn::TargetID;
 
 #[cfg(test)]
