+++ crates/notify/src/rules/rules_map.rs
@@ -14,7 +14,7 @@
 
 use crate::rules::{PatternRules, TargetIdSet};
 use hashbrown::HashMap;
-use rustfs_s3_common::EventName;
+use rustfs_s3_types::EventName;
 use rustfs_targets::arn::TargetID;
 use serde::{Deserialize, Serialize};
