+++ crates/notify/src/rules/xml_config.rs
@@ -14,7 +14,7 @@
 
 use crate::rules::pattern;
 use hashbrown::HashSet;
-use rustfs_s3_common::EventName;
+use rustfs_s3_types::EventName;
 use rustfs_targets::arn::{ARN, ArnError, TargetIDError};
 use serde::{Deserialize, Serialize};
 use std::io::Read;
