+++ crates/s3-ops/Cargo.toml
@@ -13,24 +13,22 @@
 # limitations under the License.
 
 [package]
-name = "rustfs-s3-common"
+name = "rustfs-s3-ops"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
 repository.workspace = true
 rust-version.workspace = true
 homepage.workspace = true
-description = "Common S3 definitions and metrics for RustFS."
-keywords = ["s3", "common", "rustfs"]
+description = "S3 operation enum and event mapping for RustFS."
+keywords = ["s3", "operations", "rustfs"]
 categories = ["data-structures"]
 
 [lints]
 workspace = true
 
 [dependencies]
-metrics = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
+rustfs-s3-types = { workspace = true }
 
 [lib]
 doctest = false
