+++ crates/s3-types/src/event_name.rs
@@ -315,36 +315,6 @@ impl EventName {
             mask
         }
     }
-
-    /// Returns the corresponding S3Operation if the event triggers a notification event.
-    pub fn to_s3_operation(&self) -> Option<S3Operation> {
-        match self {
-            EventName::BucketCreated => Some(S3Operation::CreateBucket),
-            EventName::BucketRemoved => Some(S3Operation::DeleteBucket),
-            EventName::ObjectAccessedGet => Some(S3Operation::GetObject),
-            EventName::ObjectAccessedGetRetention => Some(S3Operation::GetObjectRetention),
-            EventName::ObjectAccessedGetLegalHold => Some(S3Operation::GetObjectLegalHold),
-            EventName::ObjectAccessedHead => Some(S3Operation::HeadObject),
-            EventName::ObjectAccessedAttributes => Some(S3Operation::GetObjectAttributes),
-            EventName::ObjectCreatedCompleteMultipartUpload => Some(S3Operation::CompleteMultipartUpload),
-            EventName::ObjectCreatedCopy => Some(S3Operation::CopyObject),
-            EventName::ObjectCreatedPost => Some(S3Operation::PutObject),
-            EventName::ObjectCreatedPut => Some(S3Operation::PutObject),
-            EventName::ObjectCreatedPutRetention => Some(S3Operation::PutObjectRetention),
-            EventName::ObjectCreatedPutLegalHold => Some(S3Operation::PutObjectLegalHold),
-            EventName::ObjectTaggingPut => Some(S3Operation::PutObjectTagging),
-            EventName::ObjectTaggingDelete => Some(S3Operation::DeleteObjectTagging),
-            EventName::ObjectAclPut => Some(S3Operation::PutObjectAcl),
-            EventName::ObjectRemovedDelete => Some(S3Operation::DeleteObject),
-            EventName::ObjectRemovedDeleteMarkerCreated => Some(S3Operation::DeleteObject),
-            EventName::ObjectRemovedDeleteAllVersions => Some(S3Operation::DeleteObject),
-            EventName::ObjectRestorePost => Some(S3Operation::RestoreObject),
-            EventName::ObjectRemovedAbortMultipartUpload => Some(S3Operation::AbortMultipartUpload),
-            EventName::ObjectCreatedCreateMultipartUpload => Some(S3Operation::CreateMultipartUpload),
-            EventName::ObjectRemovedDeleteObjects => Some(S3Operation::DeleteObjects),
-            _ => None,
-        }
-    }
 }
 
 impl fmt::Display for EventName {
@@ -377,177 +347,6 @@ impl<'de> serde::de::Deserialize<'de> for EventName {
         let s = String::deserialize(deserializer)?;
         let s = Self::parse(&s).map_err(serde::de::Error::custom)?;
         Ok(s)
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub enum S3Operation {
-    AbortMultipartUpload,
-    CompleteMultipartUpload,
-    CopyObject,
-    CreateBucket,
-    CreateMultipartUpload,
-    DeleteBucket,
-    DeleteBucketCors,
-    DeleteBucketEncryption,
-    DeleteBucketLifecycle,
-    DeleteBucketPolicy,
-    DeleteBucketReplication,
-    DeleteBucketTagging,
-    DeleteObject,
-    DeleteObjectTagging,
-    DeleteObjects,
-    DeletePublicAccessBlock,
-    GetBucketAcl,
-    GetBucketCors,
-    GetBucketEncryption,
-    GetBucketLifecycleConfiguration,
-    GetBucketLocation,
-    GetBucketLogging,
-    GetBucketNotificationConfiguration,
-    GetBucketPolicy,
-    GetBucketPolicyStatus,
-    GetBucketReplication,
-    GetBucketTagging,
-    GetBucketVersioning,
-    GetObject,
-    GetObjectAcl,
-    GetObjectAttributes,
-    GetObjectLegalHold,
-    GetObjectLockConfiguration,
-    GetObjectRetention,
-    GetObjectTagging,
-    GetObjectTorrent,
-    GetPublicAccessBlock,
-    HeadBucket,
-    HeadObject,
-    ListBuckets,
-    ListMultipartUploads,
-    ListObjectVersions,
-    ListObjects,
-    ListObjectsV2,
-    ListParts,
-    PutBucketAcl,
-    PutBucketCors,
-    PutBucketEncryption,
-    PutBucketLifecycleConfiguration,
-    PutBucketLogging,
-    PutBucketNotificationConfiguration,
-    PutBucketPolicy,
-    PutBucketReplication,
-    PutBucketTagging,
-    PutBucketVersioning,
-    PutObject,
-    PutObjectAcl,
-    PutObjectLegalHold,
-    PutObjectLockConfiguration,
-    PutObjectRetention,
-    PutObjectTagging,
-    PutPublicAccessBlock,
-    RestoreObject,
-    SelectObjectContent,
-    UploadPart,
-    UploadPartCopy,
-}
-
-impl S3Operation {
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::AbortMultipartUpload => "s3:AbortMultipartUpload",
-            Self::CompleteMultipartUpload => "s3:CompleteMultipartUpload",
-            Self::CopyObject => "s3:CopyObject",
-            Self::CreateBucket => "s3:CreateBucket",
-            Self::CreateMultipartUpload => "s3:CreateMultipartUpload",
-            Self::DeleteBucket => "s3:DeleteBucket",
-            Self::DeleteBucketCors => "s3:DeleteBucketCors",
-            Self::DeleteBucketEncryption => "s3:DeleteBucketEncryption",
-            Self::DeleteBucketLifecycle => "s3:DeleteBucketLifecycle",
-            Self::DeleteBucketPolicy => "s3:DeleteBucketPolicy",
-            Self::DeleteBucketReplication => "s3:DeleteBucketReplication",
-            Self::DeleteBucketTagging => "s3:DeleteBucketTagging",
-            Self::DeleteObject => "s3:DeleteObject",
-            Self::DeleteObjectTagging => "s3:DeleteObjectTagging",
-            Self::DeleteObjects => "s3:DeleteObjects",
-            Self::DeletePublicAccessBlock => "s3:DeletePublicAccessBlock",
-            Self::GetBucketAcl => "s3:GetBucketAcl",
-            Self::GetBucketCors => "s3:GetBucketCors",
-            Self::GetBucketEncryption => "s3:GetBucketEncryption",
-            Self::GetBucketLifecycleConfiguration => "s3:GetBucketLifecycleConfiguration",
-            Self::GetBucketLocation => "s3:GetBucketLocation",
-            Self::GetBucketLogging => "s3:GetBucketLogging",
-            Self::GetBucketNotificationConfiguration => "s3:GetBucketNotificationConfiguration",
-            Self::GetBucketPolicy => "s3:GetBucketPolicy",
-            Self::GetBucketPolicyStatus => "s3:GetBucketPolicyStatus",
-            Self::GetBucketReplication => "s3:GetBucketReplication",
-            Self::GetBucketTagging => "s3:GetBucketTagging",
-            Self::GetBucketVersioning => "s3:GetBucketVersioning",
-            Self::GetObject => "s3:GetObject",
-            Self::GetObjectAcl => "s3:GetObjectAcl",
-            Self::GetObjectAttributes => "s3:GetObjectAttributes",
-            Self::GetObjectLegalHold => "s3:GetObjectLegalHold",
-            Self::GetObjectLockConfiguration => "s3:GetObjectLockConfiguration",
-            Self::GetObjectRetention => "s3:GetObjectRetention",
-            Self::GetObjectTagging => "s3:GetObjectTagging",
-            Self::GetObjectTorrent => "s3:GetObjectTorrent",
-            Self::GetPublicAccessBlock => "s3:GetPublicAccessBlock",
-            Self::HeadBucket => "s3:HeadBucket",
-            Self::HeadObject => "s3:HeadObject",
-            Self::ListBuckets => "s3:ListBuckets",
-            Self::ListMultipartUploads => "s3:ListMultipartUploads",
-            Self::ListObjectVersions => "s3:ListObjectVersions",
-            Self::ListObjects => "s3:ListObjects",
-            Self::ListObjectsV2 => "s3:ListObjectsV2",
-            Self::ListParts => "s3:ListParts",
-            Self::PutBucketAcl => "s3:PutBucketAcl",
-            Self::PutBucketCors => "s3:PutBucketCors",
-            Self::PutBucketEncryption => "s3:PutBucketEncryption",
-            Self::PutBucketLifecycleConfiguration => "s3:PutBucketLifecycleConfiguration",
-            Self::PutBucketLogging => "s3:PutBucketLogging",
-            Self::PutBucketNotificationConfiguration => "s3:PutBucketNotificationConfiguration",
-            Self::PutBucketPolicy => "s3:PutBucketPolicy",
-            Self::PutBucketReplication => "s3:PutBucketReplication",
-            Self::PutBucketTagging => "s3:PutBucketTagging",
-            Self::PutBucketVersioning => "s3:PutBucketVersioning",
-            Self::PutObject => "s3:PutObject",
-            Self::PutObjectAcl => "s3:PutObjectAcl",
-            Self::PutObjectLegalHold => "s3:PutObjectLegalHold",
-            Self::PutObjectLockConfiguration => "s3:PutObjectLockConfiguration",
-            Self::PutObjectRetention => "s3:PutObjectRetention",
-            Self::PutObjectTagging => "s3:PutObjectTagging",
-            Self::PutPublicAccessBlock => "s3:PutPublicAccessBlock",
-            Self::RestoreObject => "s3:RestoreObject",
-            Self::SelectObjectContent => "s3:SelectObjectContent",
-            Self::UploadPart => "s3:UploadPart",
-            Self::UploadPartCopy => "s3:UploadPartCopy",
-        }
-    }
-
-    /// Returns the corresponding EventName if the operation triggers a notification event.
-    pub fn to_event_name(self) -> Option<EventName> {
-        match self {
-            Self::CompleteMultipartUpload => Some(EventName::ObjectCreatedCompleteMultipartUpload),
-            Self::CopyObject => Some(EventName::ObjectCreatedCopy),
-            Self::CreateBucket => Some(EventName::BucketCreated),
-            Self::DeleteBucket => Some(EventName::BucketRemoved),
-            Self::DeleteObject => Some(EventName::ObjectRemovedDelete),
-            Self::DeleteObjects => Some(EventName::ObjectRemovedDeleteObjects),
-            Self::DeleteObjectTagging => Some(EventName::ObjectTaggingDelete),
-            Self::GetObject => Some(EventName::ObjectAccessedGet),
-            Self::GetObjectAttributes => Some(EventName::ObjectAccessedAttributes),
-            Self::GetObjectLegalHold => Some(EventName::ObjectAccessedGetLegalHold),
-            Self::GetObjectRetention => Some(EventName::ObjectAccessedGetRetention),
-            Self::HeadObject => Some(EventName::ObjectAccessedHead),
-            Self::PutObject => Some(EventName::ObjectCreatedPut),
-            Self::PutObjectAcl => Some(EventName::ObjectAclPut),
-            Self::PutObjectLegalHold => Some(EventName::ObjectCreatedPutLegalHold),
-            Self::PutObjectRetention => Some(EventName::ObjectCreatedPutRetention),
-            Self::PutObjectTagging => Some(EventName::ObjectTaggingPut),
-            Self::RestoreObject => Some(EventName::ObjectRestorePost),
-            Self::SelectObjectContent => Some(EventName::ObjectAccessedGet),
-            Self::AbortMultipartUpload => Some(EventName::ObjectRemovedAbortMultipartUpload),
-            Self::CreateMultipartUpload => Some(EventName::ObjectCreatedCreateMultipartUpload),
-            _ => None,
-        }
     }
 }
 
@@ -608,39 +407,6 @@ mod tests {
 
         let deserialized = serde_json::from_str::<EventName>(serialized_str);
         assert!(deserialized.is_err(), "Deserialization should fail for empty string");
-    }
-
-    #[test]
-    fn test_s3_operation_to_event_name() {
-        assert_eq!(S3Operation::PutObject.to_event_name(), Some(EventName::ObjectCreatedPut));
-        assert_eq!(S3Operation::PutObjectAcl.to_event_name(), Some(EventName::ObjectAclPut));
-        assert_eq!(S3Operation::PutObjectTagging.to_event_name(), Some(EventName::ObjectTaggingPut));
-        assert_eq!(S3Operation::DeleteObjectTagging.to_event_name(), Some(EventName::ObjectTaggingDelete));
-        assert_eq!(S3Operation::GetObject.to_event_name(), Some(EventName::ObjectAccessedGet));
-        assert_eq!(S3Operation::ListBuckets.to_event_name(), None);
-        assert_eq!(S3Operation::RestoreObject.to_event_name(), Some(EventName::ObjectRestorePost));
-        assert_eq!(S3Operation::SelectObjectContent.to_event_name(), Some(EventName::ObjectAccessedGet));
-        assert_eq!(
-            S3Operation::AbortMultipartUpload.to_event_name(),
-            Some(EventName::ObjectRemovedAbortMultipartUpload)
-        );
-    }
-
-    #[test]
-    fn test_event_name_to_s3_operation() {
-        assert_eq!(EventName::ObjectCreatedPut.to_s3_operation(), Some(S3Operation::PutObject));
-        assert_eq!(EventName::ObjectAclPut.to_s3_operation(), Some(S3Operation::PutObjectAcl));
-        assert_eq!(EventName::ObjectTaggingPut.to_s3_operation(), Some(S3Operation::PutObjectTagging));
-        assert_eq!(EventName::ObjectTaggingDelete.to_s3_operation(), Some(S3Operation::DeleteObjectTagging));
-        assert_eq!(EventName::ObjectAccessedGet.to_s3_operation(), Some(S3Operation::GetObject));
-        assert_eq!(EventName::BucketCreated.to_s3_operation(), Some(S3Operation::CreateBucket));
-        assert_eq!(EventName::Everything.to_s3_operation(), None);
-        assert_eq!(EventName::ObjectRestorePost.to_s3_operation(), Some(S3Operation::RestoreObject));
-        assert_eq!(EventName::ObjectCreatedPost.to_s3_operation(), Some(S3Operation::PutObject));
-        assert_eq!(
-            EventName::ObjectRemovedAbortMultipartUpload.to_s3_operation(),
-            Some(S3Operation::AbortMultipartUpload)
-        );
     }
 
     #[test]
