+++ crates/targets/Cargo.toml
@@ -15,7 +15,8 @@ documentation = "https://docs.rs/rustfs-targets/latest/rustfs_targets/"
 rustfs-config = { workspace = true, features = ["notify", "constants", "audit"] }
 rustfs-ecstore = { workspace = true }
 rustfs-utils = { workspace = true, features = ["notify", "tls"] }
-rustfs-s3-common = { workspace = true }
+rustfs-s3-types = { workspace = true }
+rustfs-s3-ops = { workspace = true }
 async-trait = { workspace = true }
 async-nats = { workspace = true }
 deadpool-postgres = { workspace = true }
